[Behat] Extract ProductMediaTrait from product pages
