+++ src/Sylius/Behat/Page/Admin/Product/Common/ProductMediaTrait.php
@@ -1,0 +1,185 @@
+<?php
+
+/*
+ * This file is part of the Sylius package.
+ *
+ * (c) Sylius Sp. z o.o.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Sylius\Behat\Page\Admin\Product\Common;
+
+use Behat\Mink\Element\NodeElement;
+use Behat\Mink\Exception\ElementNotFoundException;
+use Sylius\Component\Core\Model\ProductVariantInterface;
+
+trait ProductMediaTrait
+{
+    /** @var string[] */
+    private array $imageUrls = [];
+
+    public function getDefinedProductMediaElements(): array
+    {
+        return [
+            'add-image' => '[data-test-add-image]',
+            'image_subform_with_type' => '[data-test-image-subform][data-test-type="%type%"]',
+            'images' => '[data-test-images]',
+        ];
+    }
+
+    public function attachImage(string $path, ?string $type = null, ?ProductVariantInterface $productVariant = null): void
+    {
+        $this->changeTab('media');
+        $this->getElement('add-image')->click();
+
+        $this->waitForFormUpdate();
+
+        $images = $this->getElement('images');
+        $imagesSubform = $images->findAll('css', '[data-test-image-subform]');
+        $imageSubform = end($imagesSubform);
+
+        if (null !== $type) {
+            $imageSubform->fillField('Type', $type);
+        }
+
+        if (null !== $productVariant) {
+            $this->autocompleteHelper->selectByValue(
+                $this->getDriver(),
+                $imageSubform->find('css', '[data-test-product-variant]')->getXpath(),
+                $productVariant->getCode(),
+            );
+        }
+
+        $filesPath = $this->getParameter('files_path');
+        $imageSubform->find('css', '[data-test-file]')->attachFile($filesPath . $path);
+    }
+
+    public function hasLastImageAVariant(ProductVariantInterface $productVariant): bool
+    {
+        return $this->hasImageWithVariant($productVariant);
+    }
+
+    public function changeImageWithType(string $type, string $path): void
+    {
+        $filesPath = $this->getParameter('files_path');
+
+        $imageSubform = $this->getElement('image_subform_with_type', ['%type%' => $type]);
+        $imageSubform->find('css', '[data-test-file]')->attachFile($filesPath . $path);
+    }
+
+    public function removeImageWithType(string $type): void
+    {
+        $this->changeTab('media');
+
+        $imageSubform = $this->getElement('image_subform_with_type', ['%type%' => $type]);
+        $imageSubform->find('css', '[data-test-image-delete]')->click();
+        $this->waitForFormUpdate();
+    }
+
+    public function removeFirstImage(): void
+    {
+        $this->changeTab('media');
+        $firstSubform = $this->getFirstImageSubform();
+        $firstSubform->findAll('css', '[data-test-image-delete]')[0]->click();
+    }
+
+    public function hasImageWithType(string $type): bool
+    {
+        $this->changeTab('media');
+
+        try {
+            $imageSubform = $this->getElement('image_subform_with_type', ['%type%' => $type]);
+        } catch (ElementNotFoundException) {
+            return false;
+        }
+
+        $imageUrl = $imageSubform->getAttribute('data-test-image-url');
+        $this->getDriver()->visit($imageUrl);
+        $statusCode = $this->getDriver()->getStatusCode();
+        $this->getDriver()->back();
+
+        return in_array($statusCode, [200, 304], true);
+    }
+
+    public function hasImageWithVariant(ProductVariantInterface $productVariant): bool
+    {
+        $this->changeTab('media');
+        $images = $this->getElement('images');
+
+        return $images->has('css', sprintf('[data-test-product-variant="%s"]', $productVariant->getCode()));
+    }
+
+    public function countImages(): int
+    {
+        $images = $this->getElement('images');
+        $imageSubforms = $images->findAll('css', '[data-test-image-subform]');
+
+        return count($imageSubforms);
+    }
+
+    public function modifyFirstImageType(string $type): void
+    {
+        $this->changeTab('media');
+
+        $firstImageSubform = $this->getFirstImageSubform();
+
+        $firstImageSubform->find('css', 'input[data-test-type]')->setValue($type);
+    }
+
+    public function selectVariantForFirstImage(ProductVariantInterface $productVariant): void
+    {
+        $this->changeTab('media');
+
+        $imageSubform = $this->getFirstImageSubform();
+        $this->autocompleteHelper->selectByValue(
+            $this->getDriver(),
+            $imageSubform->find('css', '[data-test-product-variant]')->getXpath(),
+            $productVariant->getCode(),
+        );
+    }
+
+    public function isImageWithTypeDisplayed(string $type): bool
+    {
+        $imageElement = $this->getImageElementByType($type);
+
+        $imageUrl = $imageElement ? $imageElement->find('css', 'img')->getAttribute('src') : $this->provideImageUrlForType($type);
+        if (null === $imageElement && null === $imageUrl) {
+            return false;
+        }
+
+        $this->getDriver()->visit($imageUrl);
+        $statusCode = $this->getDriver()->getStatusCode();
+        $this->getDriver()->back();
+
+        return in_array($statusCode, [200, 304], true);
+    }
+
+    private function getFirstImageSubform(): NodeElement
+    {
+        $images = $this->getElement('images');
+        $imageSubforms = $images->findAll('css', '[data-test-image-subform]');
+
+        return reset($imageSubforms);
+    }
+
+    private function getImageElementByType(string $type): ?NodeElement
+    {
+        $images = $this->getElement('images');
+        $typeInput = $images->find('css', 'input[value="' . $type . '"]');
+
+        if (null === $typeInput) {
+            return null;
+        }
+
+        return $typeInput->getParent()->getParent()->getParent();
+    }
+
+    private function provideImageUrlForType(string $type): ?string
+    {
+        return $this->imageUrls[$type] ?? null;
+    }
+}
